refactor(discord): bootstrap の gateway イベントハンドラ配線を event-handlers.ts へ分離する
