+++ apps/discord/src/bootstrap.ts
@@ -1,8 +1,6 @@
 /* oxlint-disable max-dependencies, max-lines -- bootstrap file naturally requires many imports and lines for DI wiring */
 import { dirname, resolve } from "path";
 
-import type { DiscordAgent } from "@vicissitude/agent/discord/discord-agent";
-import { formatDiscordMessage } from "@vicissitude/agent/discord/message-formatter";
 import { GuildRouter } from "@vicissitude/agent/discord/router";
 import { McBrainManager } from "@vicissitude/agent/minecraft/brain-manager";
 import { WEB_AGENT_ID } from "@vicissitude/agent/web/web-agent";
@@ -14,7 +12,6 @@ import { createEmotionToExpressionMapper } from "@vicissitude/avatar";
 import { createGatewayApp, listenGatewayServer } from "@vicissitude/gateway/server";
 import { WsConnectionManager } from "@vicissitude/gateway/ws-handler";
 import { MemoryFactReaderImpl } from "@vicissitude/memory/fact-reader";
-import { discordDmUserIdFromScopeId, discordScopeId } from "@vicissitude/memory/namespace";
 import { ConsoleLogger } from "@vicissitude/observability/logger";
 import { METRIC, type PrometheusCollector } from "@vicissitude/observability/metrics";
 import { OllamaEmbeddingAdapter } from "@vicissitude/ollama";
@@ -37,6 +34,7 @@ import {
 } from "./bootstrap/agents.ts";
 import { loadChannelConfig } from "./bootstrap/channel-config.ts";
 import { buildCoreEnvironment, buildDiscordEnvironment } from "./bootstrap/environment.ts";
+import { setupEventHandlers } from "./bootstrap/event-handlers.ts";
 import {
 	createContextLayer,
 	createFileSessionSummaryWriter,
@@ -56,112 +54,6 @@ import {
 import { MoodNicknameService } from "./mood-nickname.ts";
 import { createPortLayout } from "./port-allocator.ts";
 import { createShutdown } from "./shutdown.ts";
-
-// ─── Event Handlers ─────────────────────────────────────────────
-
-function setupEventHandlers(deps: {
-	gateway: DiscordGateway;
-	ingestionService: MessageIngestionService;
-	metricsCollector: PrometheusCollector;
-	agents: Map<string, DiscordAgent>;
-	logger: Logger;
-	/** 信頼ユーザー (依頼者) 集合。これに含まれる authorId のメッセージに信頼マーカーを付ける */
-	trustedUserIds: string[];
-}): void {
-	const { gateway, ingestionService, metricsCollector, agents, logger, trustedUserIds } = deps;
-	gateway.onResume(() => {
-		logger.info(`[bootstrap] Discord Gateway resumed; ensuring ${agents.size} conversation loops`);
-		for (const agent of agents.values()) agent.ensurePolling();
-	});
-	gateway.onHomeChannelMessage(async (msg) => {
-		const selfUserId = gateway.getClient()?.user?.id;
-		const scopeId = msg.scopeId ?? (msg.guildId ? discordScopeId(msg.guildId) : undefined);
-		metricsCollector.incrementCounter(METRIC.DISCORD_MESSAGES_RECEIVED, {
-			guild_id: msg.guildId ?? "none",
-			channel_type: "home",
-			author_type: msg.isBot ? "bot" : "user",
-			is_thread: String(msg.isThread),
-			has_attachments: String(msg.attachments.length > 0),
-		});
-		await ingestionService.handleIncomingMessage(msg, {
-			recordConversation: true,
-		});
-		if (msg.guildId && msg.authorId !== selfUserId) {
-			const agent = scopeId ? agents.get(scopeId) : undefined;
-			if (!agent) {
-				logger.warn(`[bootstrap] no agent for guild ${msg.guildId}, message will not be processed`);
-			}
-			void agent?.send({
-				sessionKey: "home",
-				message: formatDiscordMessage(msg, { trustedUserIds }),
-				scopeId,
-				attachments: msg.attachments,
-				channelId: msg.channelId,
-				isBot: msg.isBot,
-			});
-		}
-	});
-
-	gateway.onMessage(async (msg) => {
-		const scopeId = msg.scopeId ?? (msg.guildId ? discordScopeId(msg.guildId) : undefined);
-		metricsCollector.incrementCounter(METRIC.DISCORD_MESSAGES_RECEIVED, {
-			guild_id: msg.guildId ?? "none",
-			channel_type: "mention",
-			author_type: msg.isBot ? "bot" : "user",
-			is_thread: String(msg.isThread),
-			has_attachments: String(msg.attachments.length > 0),
-		});
-		await ingestionService.handleIncomingMessage(msg);
-		if (msg.guildId && scopeId) {
-			const agent = agents.get(scopeId);
-			if (!agent) {
-				logger.warn(`[bootstrap] no agent for guild ${msg.guildId}, mention will not be processed`);
-			}
-			void agent?.send({
-				sessionKey: "mention",
-				message: formatDiscordMessage(msg, { trustedUserIds }),
-				scopeId,
-				attachments: msg.attachments,
-				channelId: msg.channelId,
-				isBot: msg.isBot,
-			});
-		}
-	});
-
-	gateway.onDirectMessage(async (msg) => {
-		const selfUserId = gateway.getClient()?.user?.id;
-		const scopeId = msg.scopeId;
-		metricsCollector.incrementCounter(METRIC.DISCORD_MESSAGES_RECEIVED, {
-			guild_id: "none",
-			channel_type: "dm",
-			author_type: msg.isBot ? "bot" : "user",
-			is_thread: String(msg.isThread),
-			has_attachments: String(msg.attachments.length > 0),
-		});
-		await ingestionService.handleIncomingMessage(msg, {
-			recordConversation: true,
-		});
-		if (!scopeId) {
-			logger.warn("[bootstrap] DM message has no scopeId, message will not be processed");
-			return;
-		}
-		if (msg.authorId === selfUserId) return;
-
-		const agent = agents.get(scopeId);
-		if (!agent) {
-			const userId = discordDmUserIdFromScopeId(scopeId) ?? "unknown";
-			logger.warn(`[bootstrap] no DM agent for user ${userId}, message will not be processed`);
-		}
-		void agent?.send({
-			sessionKey: "dm",
-			message: formatDiscordMessage(msg, { trustedUserIds }),
-			scopeId,
-			attachments: msg.attachments,
-			channelId: msg.channelId,
-			isBot: msg.isBot,
-		});
-	});
-}
 
 // ─── MCP Process Management ─────────────────────────────────────
